+++ NEWS.md
@@ -5,6 +5,7 @@
 
 * [#62](https://github.com/eblondel/zen4R/issues/62) Fix new file upload API - wrong filename spliting
 * [#63](https://github.com/eblondel/zen4R/issues/63) Fix new file upload API - case with filename with spaces
+* [#64](https://github.com/eblondel/zen4R/issues/64) Fix new file upload API - actual file upload
 
 ## **_0.5-1_**